// src/api/generated/model/admissionApplicationRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Greenfield College System API
 * Complete API Specification covering all mutations by REST.
 * OpenAPI spec version: 1.0.0
 */

export interface AdmissionApplicationRequest {
  program_public_id: string;
  first_name: string;
  last_name: string;
  email: string;
  password: string;
  phone: string;
  dob: string;
  address: string;
  emergency_contact: string;
  /** Uploaded NRC or Passport file (PDF/Image) */
  nrc_file: Blob;
  /** Uploaded High School Certificate file (PDF/Image) */
  certificate_file: Blob;
}